pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = ""
authors = [
    {name = "Kouakou-joel",email = "jojomignon5612@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
